Added common loader to tests
